requirements.txt
discord.py
beautifulsoup4
youtube_dl
python-dotenv
discord.py[voice]
pymitter
pynacl
ffmpeg
googlesearch-python
requests
youtube-search-python
pafy
